requirements.txt
boto3==1.17.88
PyYAML==5.4.1
requests==2.25.1
